+++ core/spec/jar_cli_spec.rb
@@ -44,4 +44,18 @@ describe TorqueBox::CLI::Jar do
     expect(File.exist?(lockfile)).to be false
   end
 
+  it "doesn't include excluded paths" do
+    Dir.chdir(@tmpdir) do
+      File.new("excludeme", "w")
+      Dir.mkdir("includeme")
+      File.new("includeme/excludeme", "w")
+      TorqueBox::CLI.new(%W(jar --exclude /excludeme -q --no-include-jruby
+                            --no-bundle-gems --name test.jar)).run
+      File.exist?("test.jar").should == true
+      unzip("test.jar")
+      File.exist?("app/excludeme").should == false
+      File.exist?("app/includeme/excludeme").should == true
+    end
+  end
+
 end