+++ web/ext/wunderboss-rack/org/projectodd/wunderboss/rack/RackServlet.java
@@ -40,18 +40,11 @@ public class RackServlet extends GenericServlet {
     }
 
     public final void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        RackChannel inputChannel = null;
         try {
-            inputChannel = rackApplication.getInputChannel(request.getInputStream());
             RackAdapter adapter = new ServletRackAdapter(request, response);
-            rackApplication.call(adapter, inputChannel, RackEnvironment.RACK_KEY.SERVLET_REQUEST, request);
+            rackApplication.call(adapter, RackEnvironment.RACK_KEY.SERVLET_REQUEST, request);
         } catch (Exception e) {
-            e.printStackTrace();
             throw new ServletException(e);
-        } finally {
-            if (inputChannel != null) {
-                inputChannel.close();
-            }
         }
     }